Safer reading of mmio in vfio-pci-device-open
This test program does an fwrite() direct from mmio space.  This will print
garbage, which is no real problem.  What might be a problem though is that
this fwrite() essential amounts to a memcpy() from the mmio space to
stdio's internal buffer.  Since it expects to be coming from regular
memory, that may be optimized to use exotic load and store instructions
(e.g. using vector registers).

The effects of using those on mmio space might be strange - it can result
in SIGILLs, or even crashes depending on details.

To avoid that, use a simpler plain loop to read the IO space out byte by
byte, with a volatile qualifier to discourage the compiler from doing
anything too clever.

Minor tweak of an earlier fix from Pradipta Banerjee
<pradipta.banerjee@gmail.com>.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>