Add a page fault race test
This test creates two threads, one accesses the end of the page, the
other the start.  This highlights a problem in the vfio-pci huge_fault
handler where we can race inserting a leaf PTE and a PMD entry, where
we observe a VM_FAULT_OOM error.

This can also be used with:

$ echo "func vfio_pci_mmap_huge_fault +p" > /proc/dynamic_debug/control

in order to verify that only order = 9 faults are seen in the 2MB
test and only order = 18 faults are seen in the 1G test.

NB. The huge page sizes here are specific to x86_64.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>